# api-swagger.yaml
openapi: 3.0.0
info:
  title: Caching Service API
  description: API to interact with the Caching Service for adding, retrieving, removing, and clearing cached entities.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /cache-demo/add:
    post:
      summary: Add an entity to the cache
      operationId: addItem
      requestBody:
        description: The entity to be added to the cache
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaseEntity'
      responses:
        '201':
          description: Entity added successfully
        '500':
          description: Internal server error

  /cache-demo/get/{id}:
    get:
      summary: Get an entity from the cache by ID
      operationId: getItem
      parameters:
        - name: id
          in: path
          description: ID of the entity to fetch
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Entity fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseEntity'
        '404':
          description: Entity not found
        '500':
          description: Internal server error

  /cache-demo/remove/{id}:
    delete:
      summary: Remove an entity from the cache by ID
      operationId: removeItem
      parameters:
        - name: id
          in: path
          description: ID of the entity to remove
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Entity removed successfully
        '404':
          description: Entity not found
        '500':
          description: Internal server error

  /cache-demo/clear:
    delete:
      summary: Clear all entities in the cache
      operationId: clearCache
      responses:
        '200':
          description: Cache cleared successfully
        '500':
          description: Internal server error

components:
  schemas:
    BaseEntity:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the entity